Refactoring: lets have all main CLI code together
and avoid export of CLI handler functions
